fix: type for verbatim
