+++ apps/storybook/src/stories/transfer-button/TransferButton.stories.tsx
@@ -1,7 +1,7 @@
 import { TransferButton } from "@geist/ui-react/components/transfer/transfer-button";
 import { Meta, StoryObj } from "@storybook/react";
 import { useMemo, useState } from "react";
-import { Account } from "viem";
+import type { Account } from "viem";
 import { Label } from "#components/ui/label";
 import { withMockAccount, withWagmiProvider } from "../decorators/wagmi";
 interface TransferButtonProps {
