+++ packages/domain/src/token/token.ts
@@ -1,0 +1,8 @@
+export type Token = {
+	address?: Address;
+	imageUrl?: string;
+	decimals: number;
+	name: string;
+	symbol: string;
+	type?: string;
+};
