+++ packages/domain/src/transaction/transaction.ts
@@ -1,5 +1,5 @@
-import { Address, Transaction } from "viem";
-import { Token } from "../token/token";
+import type { Transaction } from "viem";
+import type { Token } from "#token/token";
 
 // ignore other meta such as supply / volume for now
 export type TokenTransfer = Token & {
