+++ packages/ui-react/src/hooks/eas/use-get-attestations.int.test.tsx
@@ -5,7 +5,7 @@ import { useGetAttestations } from "./use-get-attestations";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { renderHook, waitFor } from "@testing-library/react";
 import React from "react";
-import { Address } from "viem";
+import type { Address } from "viem";
 
 /**
  * @vitest-environment jsdom
