+++ packages/ui-react/src/hooks/ens/efp.ts
@@ -1,4 +1,4 @@
-import { Address } from "viem";
+import type { Address } from "viem";
 
 export type AddressOrEns = Address | string;
