+++ packages/ui-react/src/hooks/ens/use-efp-api.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { Address } from "viem";
+import type { Address } from "viem";
 import {
 	AddressOrEns,
 	EfpApiOptions,
