+++ packages/ui-react/src/lib/blockscout/chain.ts
@@ -1,7 +1,7 @@
 // explorer url in chainInfo.ts are pulled from  https://github.com/blockscout/chainscout/blob/main/data/chains.json
 // https://www.blockscout.com/chains-and-projects
 
-import { Chain } from "viem";
+import type { Chain } from "viem";
 import { BLOCKSCOUT_CHAINS_BY_ID } from "./chainInfo";
 
 // TODO ensure treeshaking
