+++ packages/ui-react/src/lib/eas/easscan.ts
@@ -1,4 +1,4 @@
-import { Address, Chain } from "viem";
+import type { Address } from "viem";
 import { EAS_CONFIG_BY_CHAIN_ID } from "./config";
 
 export const getEasscanEndpoint = (chainId: number) => {
