+++ packages/ui-react/src/lib/utils/address.ts
@@ -3,7 +3,7 @@
 
 // checksum encoded
 
-import { Address, getAddress, isHex } from "viem";
+import { type Address, getAddress, isHex } from "viem";
 import { getShortHex } from "#lib/utils/hex";
 
 export type GetShortAddressReturnType = string | null;
